Combine CamundaTaskModal and CompleteTaskModal
